rptest: Adding logging of proc status during stop_node in RedpandaService.
Prior to this, core dumps triggered by sanitizers could take a long time
and cause tests to time out, but these core dumps were not visible in
the logs. This adds a way to log the proc status of each redpanda
process while it is shutting down, which can bring visibility into
whether it is core dumping.